Renames component files from index and styles to their names, and reorganize folder structure.
